+++ app/Models/Banner.php
@@ -166,7 +166,7 @@ class Banner extends Model
 
         return Cache::remember(
             'banners.' . $uuid,
-            (int)config('app.network_data_cache_ttl'),
+            config('app.network_data_cache_ttl'),
             function () use ($binId) {
                 return self::where('uuid', $binId)->with(
                     [
